Added new method for generating samples in SMRT, updated the MNIST example
